Improve name and location of type translation helpers
