+++ apis/v1beta1/util/translator/httproute_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package converter
+package translator
 
 import (
 	"testing"
